Added manuals for building, they are placed in manuals/ dir; removed obsolete/ dir; small improvements to builder.sh, crossbuilder.sh; fixed macportutil.sh, so it fix dependency paths correctly now

+++ crossbuilder.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# DOOM 2D: FOREVER BUILDER AND PACKER FOR OS X - v0.9
+# DOOM 2D: FOREVER BUILDER AND PACKER FOR OS X - v1.0
 # (c) fl0atingzero, 2022
 #
 # This version is modified for using as cross-compilation tool so you shouldn't pass arguments to it
@@ -8,7 +8,7 @@
 # dylibbundler is required
 #
 # NOTE for bundling: dependent libraries should be placed to /opt/local/lib (if you downloaded them from MacPorts)
-# That is required because dylibbundler don't work properly if libs aren't placed at their dependecty paths
+# That is required because dylibbundler don't work properly if libs aren't placed at their dependecy paths
 # NOTE 2: if you used macportsutil to obtain libaries and if you specified a different for them,
 # then macdylibbundler will probably parse dependencies since macportsutil can fix dependency paths
 
@@ -35,8 +35,16 @@ OUTDIR="/home/vlad/dfbuilddir/Doom2DF.app"
 
 # ----------------------------------------------------------------------------------------- #
 
+#flags
+# if set to 1, script will fix libraries' dependecy paths
+FLAG_L=1
+# if set to 1, script will pack bundle to DMG
+FLAG_P=1
+
+# ----------------------------------------------------------------------------------------- #
+
 echo ""
-echo "Doom 2D: Forever autobuild script for OS X - v0.9 (c) fl0atingzero, 2022"
+echo "Doom 2D: Forever autobuild script for OS X - v1.0 (c) fl0atingzero, 2022"
 echo ""
 
 
@@ -60,7 +68,6 @@ fi
 #echo "RESPATH: $RESPATH"
 #echo "FLAG_P: $FLAG_P"
 #echo "FLAG_L: $FLAG_L"
-#echo "FLAG_Y: $FLAG_Y"
 #read
 
 echo "# ----------------------------------------------------------------------------------------- #"
@@ -142,7 +149,8 @@ echo ""
 echo "Cleaning temporary files:"
 echo ""
 #rm -rf $TMPDIR/* $OUTDIR/Contents/MacOS/{link.res,ppas*}
-find $TMPDIR -mindepth 1 -delete && $OUTDIR/Contents/MacOS/{link.res,ppas*}
+find $TMPDIR -mindepth 1 -delete
+find $OUTDIR/Contents/MacOS/ \( -name "link.res" -o -name "ppas*"\) -delete
 echo "# ----------------------------------------------------------------------------------------- #"
 
 # building main
@@ -159,44 +167,55 @@ echo "Copying resources:"
 cp -rv $RESPATH/data $RESPATH/maps $RESPATH/wads $OUTDIR/Contents/Resources/
 echo "# ----------------------------------------------------------------------------------------- #"
 
-# fix library dependencies paths
-echo ""
-echo "Fixing library dependencies paths"
+# checking for "-l" flag
+if [ "$FLAG_L" == "1" ]; then
 
-cd $OUTDIR/Contents/MacOS
-echo ""
-$DYLIBBUNDLER -d $OUTDIR/Contents/libs -x Doom2DF
+    # fix library dependencies paths
+    echo ""
+    echo "Fixing library dependencies paths"
+
+    cd $OUTDIR/Contents/MacOS
+    echo ""
+    $DYLIBBUNDLER -d $OUTDIR/Contents/libs -x Doom2DF
+    echo "# ----------------------------------------------------------------------------------------- #"
+    echo ""
+    $DYLIBBUNDLER -d $OUTDIR/Contents/libs -x Doom2DF_H
 echo "# ----------------------------------------------------------------------------------------- #"
-echo ""
-$DYLIBBUNDLER -d $OUTDIR/Contents/libs -x Doom2DF_H
-echo "# ----------------------------------------------------------------------------------------- #"
+
+fi
 echo ""
 echo "main actions are done"
 
-# packing bundle to DMG
+# checking for "-p" parameter
 
-cd $OUTDIR/..
+if [ "$FLAG_P" == "1" ]; then
 
-echo ""
-echo "Creating $(pwd)/Doom2DF.dmg"
-echo ""
+    # packing bundle to DMG
 
-if [ -e $OUTDIR/../Doom2DF.root ]; then
-	echo "Doom2DF.root found, trying to delete"
-	rm -rf $OUTDIR/../Doom2DF.root
-	echo ""
+    cd $OUTDIR/..
+
+    echo ""
+    echo "Creating $(pwd)/Doom2DF.dmg"
+    echo ""
+
+    if [ -e $OUTDIR/../Doom2DF.root ]; then
+        echo "Doom2DF.root found, trying to delete"
+        rm -rf $OUTDIR/../Doom2DF.root
+        echo ""
+    fi
+
+    echo "creating new Doom2DF.root"
+    mkdir -p Doom2DF.root
+    echo ""
+    cp -rf $OUTDIR  Doom2DF.root
+    echo ""
+
+    $PACKUTIL -D -V "Doom 2D Forever" -no-pad -r -apple -file-mode 0555 -o Doom2DF.dmg Doom2DF.root
+    echo ""
+
+    echo "trying to delete unnecessary Doom2DF.root"
+    find $OUTDIR/../Doom2DF.root -delete
+
 fi
 
-echo "creating new Doom2DF.root"
-mkdir -p Doom2DF.root
 echo ""
-cp -rf $OUTDIR  Doom2DF.root
-echo ""
-
-$PACKUTIL -D -V "Doom 2D Forever" -no-pad -r -apple -file-mode 0555 -o Doom2DF.dmg Doom2DF.root
-echo ""
-
-echo "trying to delete unnecessary Doom2DF.root"
-find $OUTDIR/../Doom2DF.root -delete
-#rm -rf $OUTDIR/../Doom2DF.root
-echo ""
